remove entity directory and move to domain directory
