--- src/domain/aggregate.rs
@@ -1,2 +1,0 @@
-pub mod user;
-pub mod value_object;
