--- src/domain/aggregate/value_object.rs
@@ -1,1 +1,0 @@
-pub mod user_id;
