--- src/domain/interface/mod.rs
@@ -1,1 +1,0 @@
-pub mod user;
