+++ src/domain/mod.rs
@@ -1,2 +1,1 @@
-pub mod aggregate;
-pub mod interface;
+pub mod user;
