+++ src/domain/user/mod.rs
@@ -1,2 +1,4 @@
+pub mod model;
+pub mod prelude;
 pub mod repository;
 pub mod service;
