+++ src/domain/user/model.rs
@@ -1,4 +1,3 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
