+++ src/domain/user/prelude.rs
@@ -1,3 +1,3 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
 
-pub use super::user::Entity as User;
+pub use super::model::Entity as User;
