+++ src/domain/user/service.rs
@@ -1,7 +1,5 @@
-use crate::{
-    entities::user::user::{self, Entity as User},
-    services::user::user::CreateUserInput,
-};
+use crate::domain::user::model as user;
+use crate::services::user::user::CreateUserInput;
 use anyhow::{Error, Result};
 use sea_orm::DbErr;
