--- src/entities/user/mod.rs
@@ -1,2 +1,0 @@
-pub mod prelude;
-pub mod user;
